[issue-990] move all methods that require the Exploration object to subclasses of LandmarkFactory
