+++ src/search/landmarks/landmark_factory.cc
@@ -2,7 +2,6 @@
 
 #include "landmark_graph.h"
 
-#include "exploration.h"
 #include "util.h"
 
 #include "../option_parser.h"
@@ -20,14 +19,16 @@ using namespace std;
 
 namespace landmarks {
 LandmarkFactory::LandmarkFactory(const options::Options &opts)
-    : lm_graph_task(nullptr),
-      reasonable_orders(opts.get<bool>("reasonable_orders")),
+    : reasonable_orders(opts.get<bool>("reasonable_orders")),
       only_causal_landmarks(opts.get<bool>("only_causal_landmarks")),
       disjunctive_landmarks(opts.get<bool>("disjunctive_landmarks")),
       conjunctive_landmarks(opts.get<bool>("conjunctive_landmarks")),
-      no_orders(opts.get<bool>("no_orders")) {
+      no_orders(opts.get<bool>("no_orders")),
+      lm_graph_task(nullptr) {
 }
 /*
+  TODO: Update this comment
+
   Note: To allow reusing landmark graphs, we use the following temporary
   solution.
 
@@ -61,14 +62,9 @@ shared_ptr<LandmarkGraph> LandmarkFactory::compute_lm_graph(
     utils::Timer lm_generation_timer;
 
     TaskProxy task_proxy(*task);
-
     lm_graph = make_shared<LandmarkGraph>(task_proxy);
-
     generate_landmarks(task);
 
-    // the following replaces the old "build_lm_graph"
-    Exploration exploration(task_proxy);
-    generate(task_proxy, exploration);
     utils::g_log << "Landmarks generation time: " << lm_generation_timer << endl;
     if (lm_graph->number_of_landmarks() == 0)
         utils::g_log << "Warning! No landmarks found. Task unsolvable?" << endl;
@@ -83,45 +79,6 @@ shared_ptr<LandmarkGraph> LandmarkFactory::compute_lm_graph(
     return lm_graph;
 }
 
-void LandmarkFactory::generate(const TaskProxy &task_proxy, Exploration &exploration) {
-    if (only_causal_landmarks)
-        discard_noncausal_landmarks(task_proxy, exploration);
-    if (!disjunctive_landmarks)
-        discard_disjunctive_landmarks();
-    if (!conjunctive_landmarks)
-        discard_conjunctive_landmarks();
-    lm_graph->set_landmark_ids();
-
-    if (no_orders)
-        discard_all_orderings();
-    else if (reasonable_orders) {
-        utils::g_log << "approx. reasonable orders" << endl;
-        approximate_reasonable_orders(task_proxy, false);
-        utils::g_log << "approx. obedient reasonable orders" << endl;
-        approximate_reasonable_orders(task_proxy, true);
-    }
-    mk_acyclic_graph();
-    lm_graph->set_landmark_cost(calculate_lms_cost());
-    calc_achievers(task_proxy, exploration);
-}
-
-bool LandmarkFactory::achieves_non_conditional(const OperatorProxy &o,
-                                               const LandmarkNode *lmp) const {
-    /* Test whether the landmark is achieved by the operator unconditionally.
-    A disjunctive landmark is achieved if one of its disjuncts is achieved. */
-    assert(lmp);
-    for (EffectProxy effect: o.get_effects()) {
-        for (const FactPair &lm_fact : lmp->facts) {
-            FactProxy effect_fact = effect.get_fact();
-            if (effect_fact.get_pair() == lm_fact) {
-                if (effect.get_conditions().empty())
-                    return true;
-            }
-        }
-    }
-    return false;
-}
-
 bool LandmarkFactory::is_landmark_precondition(const OperatorProxy &op,
                                                const LandmarkNode *lmp) const {
     /* Test whether the landmark is used by the operator as a precondition.
@@ -133,105 +90,6 @@ bool LandmarkFactory::is_landmark_precondition(const OperatorProxy &op,
                 return true;
         }
     }
-    return false;
-}
-
-bool LandmarkFactory::relaxed_task_solvable(const TaskProxy &task_proxy,
-                                            Exploration &exploration,
-                                            vector<vector<int>> &lvl_var,
-                                            vector<utils::HashMap<FactPair, int>> &lvl_op,
-                                            bool level_out, const LandmarkNode *exclude, bool compute_lvl_op) const {
-    /* Test whether the relaxed planning task is solvable without achieving the propositions in
-     "exclude" (do not apply operators that would add a proposition from "exclude").
-     As a side effect, collect in lvl_var and lvl_op the earliest possible point in time
-     when a proposition / operator can be achieved / become applicable in the relaxed task.
-     */
-
-    OperatorsProxy operators = task_proxy.get_operators();
-    AxiomsProxy axioms = task_proxy.get_axioms();
-    // Initialize lvl_op and lvl_var to numeric_limits<int>::max()
-    if (compute_lvl_op) {
-        lvl_op.resize(operators.size() + axioms.size());
-        for (OperatorProxy op : operators) {
-            add_operator_and_propositions_to_list(op, lvl_op);
-        }
-        for (OperatorProxy axiom : axioms) {
-            add_operator_and_propositions_to_list(axiom, lvl_op);
-        }
-    }
-    VariablesProxy variables = task_proxy.get_variables();
-    lvl_var.resize(variables.size());
-    for (VariableProxy var : variables) {
-        lvl_var[var.get_id()].resize(var.get_domain_size(),
-                                     numeric_limits<int>::max());
-    }
-    // Extract propositions from "exclude"
-    unordered_set<int> exclude_op_ids;
-    vector<FactPair> exclude_props;
-    if (exclude) {
-        for (OperatorProxy op : operators) {
-            if (achieves_non_conditional(op, exclude))
-                exclude_op_ids.insert(op.get_id());
-        }
-        exclude_props.insert(exclude_props.end(),
-                             exclude->facts.begin(), exclude->facts.end());
-    }
-    // Do relaxed exploration
-    exploration.compute_reachability_with_excludes(
-        lvl_var, lvl_op, level_out, exclude_props, exclude_op_ids, compute_lvl_op);
-
-    // Test whether all goal propositions have a level of less than numeric_limits<int>::max()
-    for (FactProxy goal : task_proxy.get_goals())
-        if (lvl_var[goal.get_variable().get_id()][goal.get_value()] ==
-            numeric_limits<int>::max())
-            return false;
-
-    return true;
-}
-
-void LandmarkFactory::add_operator_and_propositions_to_list(const OperatorProxy &op,
-                                                            vector<utils::HashMap<FactPair, int>> &lvl_op) const {
-    int op_or_axiom_id = get_operator_or_axiom_id(op);
-    for (EffectProxy effect : op.get_effects()) {
-        lvl_op[op_or_axiom_id].emplace(effect.get_fact().get_pair(), numeric_limits<int>::max());
-    }
-}
-
-bool LandmarkFactory::is_causal_landmark(const TaskProxy &task_proxy, Exploration &exploration,
-                                         const LandmarkNode &landmark) const {
-    /* Test whether the relaxed planning task is unsolvable without using any operator
-       that has "landmark" as a precondition.
-       Similar to "relaxed_task_solvable" above.
-     */
-
-    if (landmark.in_goal)
-        return true;
-    vector<vector<int>> lvl_var;
-    vector<utils::HashMap<FactPair, int>> lvl_op;
-    // Initialize lvl_var to numeric_limits<int>::max()
-    VariablesProxy variables = task_proxy.get_variables();
-    lvl_var.resize(variables.size());
-    for (VariableProxy var : variables) {
-        lvl_var[var.get_id()].resize(var.get_domain_size(),
-                                     numeric_limits<int>::max());
-    }
-    unordered_set<int> exclude_op_ids;
-    vector<FactPair> exclude_props;
-    for (OperatorProxy op : task_proxy.get_operators()) {
-        if (is_landmark_precondition(op, &landmark)) {
-            exclude_op_ids.insert(op.get_id());
-        }
-    }
-    // Do relaxed exploration
-    exploration.compute_reachability_with_excludes(
-        lvl_var, lvl_op, true, exclude_props, exclude_op_ids, false);
-
-    // Test whether all goal propositions have a level of less than numeric_limits<int>::max()
-    for (FactProxy goal : task_proxy.get_goals())
-        if (lvl_var[goal.get_variable().get_id()][goal.get_value()] ==
-            numeric_limits<int>::max())
-            return true;
-
     return false;
 }
 
@@ -607,17 +465,6 @@ void LandmarkFactory::edge_add(LandmarkNode &from, LandmarkNode &to,
     assert(to.parents.find(&from) != to.parents.end());
 }
 
-void LandmarkFactory::discard_noncausal_landmarks(const TaskProxy &task_proxy, Exploration &exploration) {
-    int num_all_landmarks = lm_graph->number_of_landmarks();
-    lm_graph->remove_node_if(
-        [this, &task_proxy, &exploration](const LandmarkNode &node) {
-            return !is_causal_landmark(task_proxy, exploration, node);
-        });
-    int num_causal_landmarks = lm_graph->number_of_landmarks();
-    utils::g_log << "Discarded " << num_all_landmarks - num_causal_landmarks
-                 << " non-causal landmarks" << endl;
-}
-
 void LandmarkFactory::discard_disjunctive_landmarks() {
     /*
       Using disjunctive landmarks during landmark generation can be beneficial
@@ -756,31 +603,6 @@ int LandmarkFactory::calculate_lms_cost() const {
         result += lmn->min_cost;
 
     return result;
-}
-
-void LandmarkFactory::calc_achievers(const TaskProxy &task_proxy, Exploration &exploration) {
-    VariablesProxy variables = task_proxy.get_variables();
-    for (auto &lmn : lm_graph->get_nodes()) {
-        for (const FactPair &lm_fact : lmn->facts) {
-            const vector<int> &ops = lm_graph->get_operators_including_eff(lm_fact);
-            lmn->possible_achievers.insert(ops.begin(), ops.end());
-
-            if (variables[lm_fact.var].is_derived())
-                lmn->is_derived = true;
-        }
-
-        vector<vector<int>> lvl_var;
-        vector<utils::HashMap<FactPair, int>> lvl_op;
-        relaxed_task_solvable(task_proxy, exploration, lvl_var, lvl_op, true, lmn.get());
-
-        for (int op_or_axom_id : lmn->possible_achievers) {
-            OperatorProxy op = get_operator_or_axiom(task_proxy, op_or_axom_id);
-
-            if (_possibly_reaches_lm(op, lvl_var, lmn.get())) {
-                lmn->first_achievers.insert(op_or_axom_id);
-            }
-        }
-    }
 }
 
 void _add_options_to_parser(OptionParser &parser) {
