+++ src/search/landmarks/landmark_factory_h_m.cc
@@ -1,5 +1,7 @@
 #include "landmark_factory_h_m.h"
 
+#include "exploration.h"
+
 #include "../abstract_task.h"
 #include "../option_parser.h"
 #include "../plugin.h"
@@ -586,7 +588,82 @@ void LandmarkFactoryHM::initialize(const TaskProxy &task_proxy) {
     build_pm_ops(task_proxy);
 }
 
-void LandmarkFactoryHM::calc_achievers(const TaskProxy &task_proxy, Exploration &) {
+void LandmarkFactoryHM::generate(const TaskProxy &task_proxy) {
+    if (only_causal_landmarks) {
+        Exploration exploration(task_proxy);
+        discard_noncausal_landmarks(task_proxy, exploration);
+    }
+    if (!disjunctive_landmarks)
+        discard_disjunctive_landmarks();
+    if (!conjunctive_landmarks)
+        discard_conjunctive_landmarks();
+    lm_graph->set_landmark_ids();
+
+    if (no_orders)
+        discard_all_orderings();
+    else if (reasonable_orders) {
+        utils::g_log << "approx. reasonable orders" << endl;
+        approximate_reasonable_orders(task_proxy, false);
+        utils::g_log << "approx. obedient reasonable orders" << endl;
+        approximate_reasonable_orders(task_proxy, true);
+    }
+    mk_acyclic_graph();
+    lm_graph->set_landmark_cost(calculate_lms_cost());
+    calc_achievers(task_proxy);
+}
+
+void LandmarkFactoryHM::discard_noncausal_landmarks(
+    const TaskProxy &task_proxy, Exploration &exploration) {
+    int num_all_landmarks = lm_graph->number_of_landmarks();
+    lm_graph->remove_node_if(
+        [this, &task_proxy, &exploration](const LandmarkNode &node) {
+            return !is_causal_landmark(task_proxy, exploration, node);
+        });
+    int num_causal_landmarks = lm_graph->number_of_landmarks();
+    utils::g_log << "Discarded " << num_all_landmarks - num_causal_landmarks
+                 << " non-causal landmarks" << endl;
+}
+
+bool LandmarkFactoryHM::is_causal_landmark(
+    const TaskProxy &task_proxy, Exploration &exploration,
+    const LandmarkNode &landmark) const {
+    /* Test whether the relaxed planning task is unsolvable without using any operator
+       that has "landmark" as a precondition.
+       Similar to "relaxed_task_solvable" above.
+     */
+
+    if (landmark.in_goal)
+        return true;
+    vector<vector<int>> lvl_var;
+    vector<utils::HashMap<FactPair, int>> lvl_op;
+    // Initialize lvl_var to numeric_limits<int>::max()
+    VariablesProxy variables = task_proxy.get_variables();
+    lvl_var.resize(variables.size());
+    for (VariableProxy var : variables) {
+        lvl_var[var.get_id()].resize(var.get_domain_size(),
+                                     numeric_limits<int>::max());
+    }
+    unordered_set<int> exclude_op_ids;
+    vector<FactPair> exclude_props;
+    for (OperatorProxy op : task_proxy.get_operators()) {
+        if (is_landmark_precondition(op, &landmark)) {
+            exclude_op_ids.insert(op.get_id());
+        }
+    }
+    // Do relaxed exploration
+    exploration.compute_reachability_with_excludes(
+        lvl_var, lvl_op, true, exclude_props, exclude_op_ids, false);
+
+    // Test whether all goal propositions have a level of less than numeric_limits<int>::max()
+    for (FactProxy goal : task_proxy.get_goals())
+        if (lvl_var[goal.get_variable().get_id()][goal.get_value()] ==
+            numeric_limits<int>::max())
+            return true;
+
+    return false;
+}
+
+void LandmarkFactoryHM::calc_achievers(const TaskProxy &task_proxy) {
     utils::g_log << "Calculating achievers." << endl;
 
     OperatorsProxy operators = task_proxy.get_operators();
